+++ src/nougaro.py
@@ -114,7 +114,9 @@ def run(
     # run the code (interpreter)
     if work_dir is None:
         work_dir = noug_dir
-    interpreter = src.runtime.interpreter.Interpreter(run, noug_dir, new_args_strings, work_dir, file_name)
+    interpreter = src.runtime.interpreter.Interpreter(
+        run, noug_dir, new_args_strings, work_dir, lexer_metas, file_name
+    )
     if use_context is None:
         context = Context('<program>', tokens[0].pos_start, None)  # create the context of the interpreter
         # don't forget to change the context symbol table to the global symbol table