+++ src/runtime/values/basevalues/value.py
@@ -246,8 +246,8 @@ class Value:
         return None, self.illegal_operation()
 
     def execute(self, args: list[Value], interpreter_: type[Interpreter], run: RunFunction, noug_dir: str,
-                exec_from: str = "<invalid>", use_context: Context | None = None, cli_args: list[String] | None = None,
-                work_dir: str | None = None):
+                lexer_metas: dict[str, str | bool], exec_from: str = "<invalid>", use_context: Context | None = None,
+                cli_args: list[String] | None = None, work_dir: str | None = None):
         """Execute the function.
         Returns a result"""
         return RTResult().failure(self.illegal_operation())